+++ src/SensorSpritesExtension.ts
@@ -54,7 +54,7 @@ export class SensorSpritesExtension extends UIBaseExtension {
     }
 
     onToolbarCreated() {
-        this._createToolbarUI('iot-sensor-sprites-btn', 'IoT Sensor Sprites', 'https://img.icons8.com/ios-filled/50/000000/iot-sensor.png'); // <a href="https://icons8.com/icon/61307/iot-sensor">IoT Sensor icon by Icons8</a>
+        this.createToolbarButton('iot-sensor-sprites-btn', 'IoT Sensor Sprites', 'https://img.icons8.com/ios-filled/50/000000/iot-sensor.png'); // <a href="https://icons8.com/icon/61307/iot-sensor">IoT Sensor icon by Icons8</a>
     }
 
     _onSpriteClicked(ev: any) {
@@ -72,7 +72,7 @@ export class SensorSpritesExtension extends UIBaseExtension {
         viewableData.spriteSize = 32;
         this._dbIdToSensorId.clear();
         let dbid = 1000000;
-        for (const [sensorId, sensor] of this.dataView.sensors.entries()) {
+        for (const [sensorId, sensor] of this.dataView.getSensors().entries()) {
             this._dbIdToSensorId.set(dbid, sensorId);
             const { x, y, z } = sensor.location;
             const style = this._style as Autodesk.DataVisualization.Core.ViewableStyle;